LayoutTests/crypto/subtle/rsassa-pkcs1-v1_5-import-key-verify-sha1-expected.txt

Test verification with RSASSA-PKCS1-v1_5 SHA-1 using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS verified is true
PASS successfullyParsed is true

TEST COMPLETE
